Merge bk://linux-input.bkbits.net/linux-input into home.transmeta.com:/home/torvalds/v2.5/linux
2002/10/11 10:05:46-07:00 mingo
[PATCH] sched-2.5.41-A0

This fixes the scheduler's migration code to not disable preemption.  It
also fixes the bug that was hidden by the broken disable-preempt change:
the migration thread did not kick idle CPUs if a task is migrated to
them, which causes a hung boot when ksoftirqds are started.  It was pure
luck it worked until now, it was broken pretty much from day 1 on.

BKrev: 3da707d32ftcZC4VA_sRkrQTdtyq4g