[PATCH] A basic NFSv4 client for 2.5.x
This is a nontrivial change to the NFS client.

This patch does for the async WRITE and COMMIT paths what patch 14
did for the async READ path, by changing 'struct nfs_write_data'
so that the v2- and v3-specific parts are moved into a private area,
with room for a v4-specific part in parallel.  None of the logic is
changed.

BKrev: 3da765dcR8OUyWu0vqfrdHldizhStw