[PATCH] kNFSd: New routine exp_pseudoroot() to find 'root' filehandle for nfsv4
NFSv4 has a concept of a 'root' filehandle (from which all other can
be found via LOOKUP, so no need for mountd).

This patch defines exp_pseudoroot(), which is used to set the filehandle
for the root of the pseudofs. The root is defined to be the export
point with fsid == 0.

BKrev: 3da78a34CqCKSBKgcj-GfdNrlO8O5w