[PATCH] kNFSd: Initialial caching infrastructure for RPC authentication caches
This patch provides a "virtual class" for defining caches
that make user-space information available in the kernel
It is intended for RPC services or clients that need user-space
support for authentication.

As yet, support for userspace interaction isn't included as I want
that to be able to have separate review.

BKrev: 3da78b57fi_JKk8TukGVt2irUa2UYw