[PATCH] kNFSd: decode symlink inplace to avoid modifying request
Rather than possibly modifying a request (as
xdr_decode_string can) we use xdr_decode_string_inplace
to symlink contents.  This string may not be nul terminated
(though it often is) so in the 'unlikely' event that is isn't
nul terminated, we copy it into a kmalloced space first.

It might be nice if vfs_symlink took a length, but then every
filesystem would have to as well....

BKrev: 3da78b7bwjZphz3W9gPfe59e5tvc4g