[PATCH] x86-64 - new memory map handling
New e820 memory map handling for x86-64.  Move it all to a new file and
clean it up a lot.  Add some simple allocator functions to deal with
holey e820 mappings cleanly A lot of this is preparation for NUMA (which
works in 2.4, but is not ported to 2.5 yet)

BKrev: 3da8b8a9YsgTSjMgny6VVTOxgRgRRA