[PATCH] fix disk IO stats for 512-byte IOs
If you're peforming 512-byte sized IOs.  With, say,

	dd of=/dev/raw/raw1 bs=512

then the `pgpgin' and `pgpgout' accounting just sits on zero.

This is because it counts in kbytes, and 512/1024 is zero.

So change it to count sectors, and divide that by two when we report it
to userspace.

BKrev: 3daa08baCBA3VLlmVr3GzslD7Z3IJw