XFS: Rework dev_t and linux inode handling
This is a two fold change, first it moves the translation between
linux dev_t and kdev_t up the call stack in xfs and makes the bulk
of xfs work in terms of its on disk dev_t format.  It also cleans
up a few related chunks of code.

The other part of the change reworks how we keep the linux inode
contents and the xfs inode fields in sync. A number of places where
we resynced the two have been removed, these were basically
replicating work elsewhere in the filesystem.   We now also ensure
that the inode fields are filled in before calling unlock_new_inode -
there used to be a window.

Finally all the code which hooks together the linux inode and the xfs
inode is brought together as a more coherent whole rather than being
cattered around the inode create path. Most calls to revalidate the
linux inode from the xfs inode are removed.

Modid: 2.5.x-xfs:slinx:128899a 10/02/02

BKrev: 3dab2bc0yt4ewSDH3JpqkEtqQYnc3Q