XFS: Fix a couple of nasty log problems
When a transaction crosses multiple iclogs, the async transaction code
needs to force the log up until the last iclog.  We need to record this
lsn in the transaction so we can do a log flush on it.

Secondly, there was a sleep/wakeup pair between flushing the log and
log write completions which was a) incorrect, and  b) no longer needed.
This could result in early wakeups of threads waiting for log flushes.

Modid: 2.5.x-xfs:slinx:129778a

BKrev: 3dab36a0LXBLeNFTcvYX6OohhR2duw