XFS: Switch from iget_locked to ilookup in vn_get
We only want to get the inode if it actually is present in core.  The
new ilookup function allows us to do this easily instead of having to
tear down the wrongly allocated inode again if it wasn't in memory.

Modid: 2.5.x-xfs:slinx:129883a

BKrev: 3dab3759Qlz9dLaT7wwiqWHxX3USnA