Merge mulgrave.(none):/home/jejb/BK/scsi-misc-2.5 into mulgrave.(none):/home/jejb/BK/scsi-for-linus-2.5
2002/10/17 21:57:38-05:00 andmike
[PATCH] scsi_debug new scan fix

If you use scsi_debug the patch below fixes a problem that has existed
since the updated scan code was merged in. scsi_debug previously assumed
that the scsi_device used to probe and the device permanently added to the
host_queue would be the same. This caused scsi_debug to allocate some
internal data and key it off the scsi_device pointer for later use.  This
resulted in scsi_debug failing all IO post scanning. This patch corrects
this problem.

Note:
	Douglas Gilbert is the maintainer of this driver.
	dougg@gear.torque.net
	http://www.torque.net/sg/

	During Douglas Gilbert's time-off he connects when he can so it
	maybe a bit until he can address this.

	In the interim this patch makes scsi_debug functional.


The full patch is available at:
http://www-124.ibm.com/storageio/patches/2.5/scsi-debug
-andmike
--
Michael Anderson
andmike@us.ibm.com

 scsi_debug.c |   59 +++++++++++++++++++++++++++++++----------------------------
 1 files changed, 31 insertions(+), 28 deletions(-)

2002/10/17 21:57:26-05:00 patmans
[PATCH] Re: [PATCH] SCSI-2 LUN setting consolidation

This patch consolidates the setting of the LUN in byte 1 of the SCSI
command block for SCSI-2 and lower devices.

This is needed for multi-path IO (some devices can actually have differnt
LUN values for each path), but is also a nice clean up of the code.

sg.c was able to inhibit setting the value, this removes that capability -
if it is really needed it can be black listed in the device_list[] flags,
or sg could set a similiar flag.

This is patched against the latest bk (as of some time on Oct 16), and
patches clean against 2.5.43.

Tests were run with different adapters and various disks and a tape drive,
I have no SCSI-2 multi-lun devices, so the main verifcation was done using
a modified scsi_debug.c driver to emulate a multi-lun SCSI-2 device.

 drivers/scsi/osst.c       |    2 --
 drivers/scsi/scsi.c       |    6 ++++++
 drivers/scsi/scsi_error.c |   12 +++++-------
 drivers/scsi/scsi_ioctl.c |   16 ++++------------
 drivers/scsi/scsi_scan.c  |   31 ++++---------------------------
 drivers/scsi/sd.c         |   22 +++++-----------------
 drivers/scsi/sg.c         |    4 ----
 drivers/scsi/sr.c         |   17 ++---------------
 drivers/scsi/sr_ioctl.c   |   20 +-------------------
 drivers/scsi/sr_vendor.c  |   18 +++---------------
 drivers/scsi/st.c         |    2 --
 include/scsi/sg.h         |    2 +-
 12 files changed, 31 insertions(+), 121 deletions(-)

BKrev: 3daf80dcVXJ5LeiFe9oGZLcyN_ZVCw