Merge bk://ppc.bkbits.net/for-linus-ppc into home.transmeta.com:/home/torvalds/v2.5/linux
2002/10/18 23:00:32+10:00 paulus
PPC32: allow for host bridges that have a type field in the addr word.

Some host bridges that use the indirect (addr and data registers)
method of access have a field that needs to be set to 0 for accesses
to the primary bus and 1 for accesses to subordinate buses.  This
change allows for that.

2002/10/18 22:50:54+10:00 paulus
PPC32: make do_div handle all 64 bits of the dividend (not just 32).

We do this with an out-of-line assembler function, __div64_32,
for the case when the high word of the dividend is non-zero.

2002/10/18 22:46:50+10:00 paulus
PPC32: Add some extra kernel debugging options.

2002/10/18 22:43:33+10:00 danc
PPC32: Add support for the SBS Palomar 4 embedded board.

2002/10/18 22:40:41+10:00 cort
PPC32: Fixes for Force PowerCore boards.

2002/10/18 22:38:01+10:00 trini
PPC32: improve the memory size detection for PReP machines.

2002/10/18 22:36:25+10:00 paulus
PPC32: fix a typo in mpc10x.h.

2002/10/18 22:35:02+10:00 paulus
PPC32: Update boot wrapper Makefiles.

We can't now use make -C in conjunction with Rules.make because
the rules for making .o files break if we do.  This updates the
boot wrapper Makefiles so that we don't use make -C, instead we
use call descend plus various other new features in Rules.make.

2002/10/18 22:30:03+10:00 paulus
PPC32: export open/read/lseek/close for the sake of ALSA modules.

2002/10/18 22:28:28+10:00 paulus
PPC32: make xchg take volatile *, remove duplicate xchg_u32 defn.

Previously we had xchg_u32 both inline in system.h and out-of-line
in arch/ppc/kernel/misc.S, and it was exported for good measure.
We only need the inline definition.

2002/10/18 22:24:07+10:00 trini
PPC32: Update section names to .foo.text/.foo.data form.

BKrev: 3db03b5dfrzGnWwWwEgDz5Nr_v-EEg