[PATCH] radix_tree_gang_lookup fix
When performing lookups against very sparse trees
radix_tree_gang_lookup fails to find nodes "far" to the right of the
start point.  Because it only understands sparseness in the leaf nodes,
not the intermediate nodes.

Nobody noticed this because all callers are incrementing the start
index as they walk the tree.

Change it to terminate the search when it really has inspected the last
possible node for the current tree's height.

BKrev: 3dbeb5cf71ROnE7np-nkzvozrJZ8uQ