Merge bk://linuxusb.bkbits.net/pnp-2.5 into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2002/10/29 09:45:25-08:00 greg
Merge kroah.com:/home/greg/linux/BK/bleeding_edge-2.5
into kroah.com:/home/greg/linux/BK/pnp-2.5

2002/10/28 21:30:10-08:00 greg
merge

2002/10/24 00:27:01-07:00 ambx1
[PATCH] update PnP layer to driver model changes - 2.5.44 (4/4)

Updates to the driver model changes.  This should fix a potential panic.

2002/10/24 00:26:22-07:00 ambx1
[PATCH] Convert CS4236B driver - 2.5.44 (3/4)

This patch converts the CS4236B sound card driver to the new PnP APIs.  Also it
makes pnp_driver_register return the number of matches during the driver add.
This should serve as a sample driver, along with the serial and parport_pc.

2002/10/24 00:25:49-07:00 ambx1
[PATCH] PnPBIOS changes - 2.5.44 (2/4)

This patch adds compatible PnP ID support to the PnPBIOS protocol.  None of my
test systems take advantage of this feature but it is included in the
specifications so it makes sense to support it.  If anyone does get a compatible
ID listed for the PnPBIOS I'd be interested to hear about it (if more than 1 id
is listed when viewing the driverfs file 'id' within the PnPBIOS protocol).  Also
it fixes the dma and mem resource problem.

2002/10/24 00:25:27-07:00 ambx1
[PATCH] PnP cleanups and resource changes - 2.5.44 (1/4)

This patch fixes a number of things pointed out by Arne Thomassen.  Also it
makes a few changes to the resource checking functions in that they now check to
make sure that resources do not conflict within the same device instead of only
other devices.  Although it is rare for this to be a factor it's nice to be able
to deal with such situations properly.

2002/10/21 11:39:24-07:00 ambx1
[PATCH] PnP Rewrite Fixes - 2.5.44

This patch addresses a few minor issues for the Linux Plug and Play Rewrite.  It
is against 2.5.44.

They are as follows.

1.) fix Config.in file - from Adrian Bunk and Roman Zippel
2.) if unable to activate a device the match should fail.  This can be done now
that the driver model matching bug has been corrected.
3.) move compat.c to isapnp directory and fix everything accordingly - suggested
by Stelian Pop.  This fixes a compile error if ISAPNP is disabled.
4.) fix a typo in pnp.h - patch from Skip Ford

Please Apply,
Adam

BKrev: 3dbeec5cMPJbEi55qN-EXvKFa7TpMA