Merge bk://ldm.bkbits.net/linux-2.5-kobject into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2002/10/30 12:50:14-08:00 mochel
Merge osdl.org:/home/mochel/src/kernel/devel/linux-2.5-virgin
into osdl.org:/home/mochel/src/kernel/devel/linux-2.5-kobject

2002/10/30 12:41:43-08:00 mochel
kobjects: add array of default attributes to subsystems, and create on registration.

struct subsystem may now contain a pointer to a NULL-terminated array of
default attributes to be exported when an object is registered with the subsystem.
kobject registration will check the return values of the directory creation and
the creation of each file, and handle it appropriately.

The documentation has also been updated.

2002/10/30 12:27:36-08:00 mochel
sysfs: kill struct sysfs_dir.

Previously, sysfs read() and write() calls looked for sysfs_ops in the struct
sysfs_dir, in the kobject. Since objects belong to a subsystem, and is a member
of a group of like devices, the sysfs_ops have been moved to struct subsystem,
and are referenced from there.

The only remaining member of struct sysfs_dir is the dentry of the object's
directory. That is moved out of the dir struct and directly into struct kobject.
That saves us 4 bytes/object.

All of the sysfs functions that referenced the struct have been changed to just
reference the dentry.

2002/10/30 11:47:41-08:00 mochel
Introduce struct subsystem.

A struct subsystem is basically a collection of objects of a certain type,
and some callbacks to operate on objects of that type.

subsystems contain embedded kobjects themselves, and have a similar set of
library routines that kobjects do, which are mostly just wrappers for the
correlating kobject routines.

kobjects are inserted in depth-first order into their subsystem's list of
objects. Orphan kobjects are also given foster parents that point to their
subsystem. This provides a bit more rigidity in the hierarchy, and disallows
any orphan kobjects.

When an object is unregistered, it is removed from its subsystem's list. When
the objects' refcount hits 0, the subsystem's ->release() callback is called.

Documentation describing the objects and the interfaces has also been added.

2002/10/30 01:26:38-08:00 mochel
sysfs: make symlinks easier.

It's now

int sysfs_create_link(struct kobject * kobj, struct kobject * target, char * name)

So, the caller doesn't have to determine the path of the target nor the depth of
the object we're creating the symlink for; it's all taken care of.


2002/10/29 23:14:59-08:00 mochel
Merge osdl.org:/home/mochel/src/kernel/devel/linux-2.5-virgin
into osdl.org:/home/mochel/src/kernel/devel/linux-2.5-kobject

2002/10/29 18:26:01-08:00 mochel
sysfs: marry api with struct kobject.

This works on obviating the need for a separate data type to describe a sysfs
directory (which was renamed from struct driver_dir_entry to struct sysfs_dir).

All sysfs creation and removal functions now take a struct kobject, instead
of a struct sysfs_dir. This kobject is embedded in ->d_fsdata of the directory.

sysfs_create_dir() takes only 1 parameter now: the object that we're creating
the directory for. The parent dentry is derived by looking at the object's
parent.

sysfs_create_file() takes the object as the first parameter, and the attribute
as the second, which makes more sense from an API perspective.

sysfs_remove_file() now takes an attribute as a second parameter, to be
consistent with the creation function.

sysfs_remove_link() is created, which is basically the old sysfs_remove_file().
(symlinks don't have an attribute associated with them; only a name, which was
prohibiting the previous change).

open() and close() look for a kobject now, and do refcounting directly on the
object. Because of that, we don't need the ->open() and ->close() callbacks
in struct sysfs_ops, so they've been removed.

read() and write() also now look for a kobject now.

The comments have been updated, too.

2002/10/29 17:54:07-08:00 mochel
sysfs: convert sysfs to use more functions from fs/libfs.c

2002/10/29 17:19:51-08:00 mochel
introduce struct kobject: simple, generic object for embedding in other structures.

This is not meant to be fancy; just something simple for which we can
control the refcount and other common functionality using common code.

The basic operations for registration and reference count manipulation are
included.

BKrev: 3dc04d3eXY6fNMuDevWbl9uD590e-A